adapter: remove ConnClient
The ConnClient was meant to safely handle the allocation and
deallocation of connection IDs, but exhibited a race condition (#18996)
that was fixed in 8ebbc08. Since that refactor, it no longer carries its
weight. Replace the ConnClient with a new_conn_id method on Client, and
then allow converting a Client directly into a SessionClient without the
intermediate ConnClient.