zippy: Introduce true upsert operations
Previously, Zippy was either inserting fresh records or
deleting existing ones. With this modification, some operations
will be true upsert operations that are modifying existing records
by changing their "pad" field.

The size of the pad follows a Zipfian distribution so that
both small and large records are represented.

Furthermore, 75% of Kafka topics created will be upsert topics,
to match with the overall focus of development.