Analyze number of attributes for auto_index_selects
This commit introduces a syntactic analysis in sqllogictest so that we can obtain the number of attributes of a given SELECT statement, except if there are wildcards. This analysis is important to avoid, whenever possible, the problem of column ambiguity leading to failed indexed view creations for the corresponding SELECT. Additionally, the commit also ensures that if expected column names are given in a test, then these are used to create the corresponding indexed view for the SELECT statement.