Handle ORDER BY in auto_index_selects
This commit extends the implementation of the auto_index_selects option of sqllogictest to better handle ORDER BY clauses in SELECT queries. When included in a view, the ORDER BY clause can be ignored. So it needs to be converted into an equivalent ORDER BY clause when querying the view. Given that we can rename the output schema to avoid column ambiguities, the converted ORDER BY clause is written positionally, for which we need to match the corresponding expressions in the projection of the original query.