sql: improve errors when creating an SSH-tunneled kafka connection
While waiting for a fix for #18491, we added a retry loop to connection
validation for Kafka connections involving an SSH tunnel. Now that the
issue is fixed, remove the retry loop.

We've observed that the retry loop can make connection errors worse.
Some SSH servers are configured to reject repeated attempts to connect
with a "connection refused" error; since the retry loop returns only the
last error, this can result in a more accurate "permission denied" error
becoming a "connection refused" error by the time it is reported to the
user.