adapter: Fix default privilege stash migrations (#19820)
7e91d54e342a9cae3756b2a55db71202ea8678b5 Added a new stash collection
for default privileges, and initialized it with a single USAGE privilege
for all types. However, the initialization logic is only run on
uninitialized stashes. As a result, stashes that were upgraded never got
the single row with the USAGE privilege.

This commit updates the upgrade logic so that the USAGE privilege is
added during stash upgrades.

Fixes #19811