persist: fix bug where apply_diff wasn't updating some State fields
In particular, it was missing applier_version, walltime_ms, and
hostname. Categorically avoid this sort of bug in the future by
deconstructing StateDiff so new fields become compile errors.

We _could_ write a regression test for this, but it wouldn't catch any
later bugs of the same category (unless we did the same destructuring
trick), so it's not worth it. OTOH, running [this assertion] in release
would be expensive but would have caught this bug immediately.

[this assertion]: https://github.com/MaterializeInc/materialize/blob/f427dc87d88dc5699b5c3d259be772df0cbad92c/src/persist-client/src/internal/apply.rs#L438-L445