compute-client: message size metrics
This commit refactors the existing message size metrics collected by the
compute controller to hopefully make them more useful. The following
changes are made:

 * Change the metric types from histogram to counter
 * Rename the `instance` label to `instance_id`
 * Rename the `replica` label to `replica_id`
 * Add new `command_type`/`response_type` labels

The label renamings are to reduce ambiguity in the naming, and to bring
these labels more in line with the `worker_id` label we use on the
replica side.

The change from histogram to counter is to reduce the metrics'
cardinalities. The assumption here is that a histogram doesn't provide
very much use over a counter. We trade the cardinality budget for an
additonal type label that allows us to split counts by message type.