Exit the offset search when we've passed the expected offset
In get_latest_ts, we rely on finding a PartitionEOF error from rdkafka
to know when to stop looking for our own writes in the progress topic.
However, rdkafka does not signal this error reliably, especially when
the partition is also receiving writes. (And it's an inherently
best-effort mechanism; partition's don't really have an end in the usual
sense.)

Instead of waiting to catch up to the end, we instead try and catch up
to whatever the high watermark was when we started up. A safety
argument:
- We initialize transactions before reading the progress records. At
  this point, any writes from a previous version of this sink should be
  visible, so the offset we fetch should be larger than any
  previously-written progress record and thus picked up by our scan.
- If another sink spins up and fences us out, we may not see the latest
  progress record... but since we're fenced out, we'll be unable to act
  on our stale information.

And anyways the old EOF approach didn't really have any guarantees
attached, so we shouldn't be at risk of making this worse in that
respect...