sql: rewrite functions in FROM position
In support of permitting scalar functions in FROM, include the
AST transformations of functions when they're in the FROM
position.

Because these transformations rewrite functions in terms of
exprs, we rewrite them as exprs in subqueries.

However, we black aggregate functions from being rewritten in
this position to mirror Postgres' prohibition; this is the right
approach if we ever move away from AST rewrites and implement these
as properly executed functions.