Regression test for #19610
This commit introduces a regression test for the correctness problem observed in issue #19610. Correctness was compromised due to nondeterminism associated to the instantiation of the hash function across workers as well as reuse of the hash function in a single worker. The test forces consolidation followed by a monotonicity assertion via a monotonic one-shot SELECT, which will trick an incorrect hash exchange to produce non-monotonic output from consolidation. The test employs cursors to avoid that retries would chance upon the correct result, and also introduces repetitions to counteract (un)lucky strikes in the hash function instantiation. The commit also makes minor adjustments to code comments.