Add two-level strategy for auto_index_selects
This commit introduces a two-level strategy for view-based execution of SELECTs under auto_index_selects in sqllogictest. The first level attempts view-based execution without any analysis of the SQL syntax, which can maximize ORDER BY clause reuse albeit being exposed to column ambiguity problems. The second level is activated upon an inconsistent outcome in the first level. Here, analysis of the number of attributes and a corresponding ORDER BY rewrite is attempted before producing the final outcome. The ORDER BY rewrite is enriched in this commit with addition of expressions that are not found in the projection list to the SELECT in the view definition, which are then projected away when querying the view.