Revisit single-time dataflow test
This commit revisits the test that determines whether a dataflow is evaluated over a single time. The test previously did not consider a corner case where the as_of would contain the maximum integer value for the timestamp and the until would be empty. In this case, it was incorrectly determined that the dataflow was not single-time, as illustrated in #19290. A fix is provided along with a regression test.