bin/pydoc: fix crash
Due to a bug in Python [0], pydoc crashes on empty __init__.py files.
Fix this by adding a copyright header to the empty cloudtest __init__.py
file.

Also put the materialize/setup.py file behind an import guard and remove
a copyright header that does not apply to the file.