adapter: Replace role attributes with system priv (#20169)
This commit replaces the usage of role attributes with system
privileges. When a statement is executed, depending on the specific
statement, the user's system privileges may be checked to determine if
the user is allowed to execute that statement. Role attributes will
never be checked for privileges.

Role attributes still exist, and they can still be modified, but they
have no effect. Role attributes will be removed in a later backwards
incompatible commit.

Works towards resolving #19997