adapter: connection validation on creation and on demand (#19487)
This PR adds a new feature giving the ability to users to validate
`CONNECTION` objects on creation and on demand. This accomplished by two
modifications in the SQL we accept:

**Validation on creation:** All `CREATE CONNECTION` statements can now
be modified with the familiar "WITH" options. For now the only option
that can be specified is the `VALIDATE` option whose value must be
either `true` or `false`. When the option is not specified a sensible
default is picked depending on the type of connection. The default
values can be seen in the following table:

| Type            | Validate by default |
|-----------------|---------------------|
| Postgres        | Yes                 |
| Kafka           | Yes                 |
| Schema Registry | Yes                 |
| AWS             | No                  |
| AWS Privatelink | No                  |
| SSH             | No                  |

**Validation on demand:** All `CONNECTION` objects can be validated on
demand using a new `VALIDATE CONNECTION <name>` statement. This query
will block until the validation completes and will either successfully
finish, returning no rows, or it will return SQL error with details
about what went wrong.

**NOTE:** This PR does *NOT* actually implement validation for `SSH`,
`AWS`, and `AWS PRIVATELINK` connections. This will be done in a
separate PR to leave this one focused on the validation infrastructure.
`KAFKA`, `POSTGRES`, and `CSR` connections are validated as part of the
PR since it has trivial to do. Of course if a `KAFKA`/`POSTGRES`/`CSR`
connection *does* use ssh/privatelink then those are implicitly
validated too.

---------

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>