adapter: automated exhausitive function tests
Add a test that executes most scalar functions with all combinations of
their interesting datums. Errors are acceptable at any point, but panics
will cause the test to fail (well, panic). Additionally assert the
returned datum is of the expected type if we get that far.

This found many problems, all fixed here. All of the Int64 -> Int32
changes now bring us inline with postgres, so those are good anyway.
Changing the arguments to int32s allowed the panics found here (things
like `i64::MIN - 1`) to now not panic because upcasting the i32 to an
i64 makes those operations safe. Add some arbitrary limitations to some
functions that can allocate lots of memory (instead of crashing
or...allocating a huge amount of memory).