storage/pg: move invariant comparing snapshot and resume_lsn
Prior to #20329, RewindRequests' snapshot_lsn was the the
consistent_point returned from CREATE_REPLICATION_SLOT. We then
asserted that the resume_lsn was less than or equal to this value.

However, in #20329, we started subtracting 1 from the
consistent_point but did not change the assertion's equality.

In a CI[1] run using #20329, I observed a case where the consistent
point was equal to the resume LSN, and our subtraction of one
caused the assertion to fail.

That this means is that the temporary replication slot used for
snapshotting had a consistent point equal to the
`confirmed_flush_lsn`; often times starting a new replication slot
advances the WAL, meaning that its consistent point will be
beyond the `confirmed_flush_lsn`, but nothing I have found
necessitates that to be the case. It also seems like this check
simply moves us back to the state we were in before #20329, so
should be somewhat uncontroversial.

[1]: https://buildkite.com/materialize/tests/builds/58992#01892d0c-15f3-470d-807f-1fbe8e3d254a