Add indexed view scenarios to feature benchmark
This commit adds three additional Dataflow scenarios to the feature
benchmark corresponding to indexed-view versions of MinMax, GroupBy,
and DeltaJoin. The dataflows created in these new scenarios utilize
operators used to maintain indexes or materialized views as opposed
to monotonic ones used for one-shot SELECTs in the case of MinMax and
GroupBy. As for DeltaJoin, the new scenario creates the base data with
a table and not a constant view, under the assumption that not
emitting the empty frontier implies the need to maintain more memory
active. All of the scenarios introduced appear to require more memory
to be maintained, and are thus useful to increase variety in support
of detecting memory regressions.