Introduce warning outcome in sqllogictest
This commit introduces a warning outcome in sqllogictest. A warning is
printed like an error, depending on the verbosity, but does not cause
the run to fail. In addition, the commit prepares for, but does not yet
realize, the use of the warning outcome in inconsistent view outcomes.
The latter outcomes can be turned into warnings in selected situations,
which are to be determined by the future implementation of the function
should_warn.