Fix AS OF handling in auto_index_selects
This commit fixes the handling of the AS OF clause in sqllogictest's
--auto-index-selects option. Previously, the AS OF clause of a SELECT
statement was ignored. With the commit, the AS OF clause is now
considered in the view-based execution by getting attached to the query
on the indexed view created for the SELECT. This fix enables
temporal.slt and as_of.slt to pass without any inconsistent view
outcome failures.