expr: allow using current_schema when cluster is invalid
Since 6a93648, we can execute `SELECT current_schemas()` without a valid
cluster. Unfortunately the same was not true for `SELECT
current_schema`, since it was implemented as `SELECT
current_schemas()[1]`, and we can't execute array subscripts without a
valid cluster.

Implement `current_schema` with a dedicated unmaterializable function,
so that it does not require an array subscript operation and therefore
does not require a valid cluster.

This fixes an issue with PopSQL, which executes `SELECT current_schema`
to validate a connection. We don't want this connection validation to
depend on the active cluster existing, as otherwise you can't use PopSQL
to run `CREATE CLUSTER default` if you have dropped the default cluster.